+++ ICSharpCode.Decompiler/TypeSystem/Implementation/KnownAttributes.cs
@@ -44,6 +44,7 @@ namespace ICSharpCode.Decompiler.TypeSystem
 		NullablePublicOnly,
 		Conditional,
 		Obsolete,
+		Embedded,
 		IsReadOnly,
 		SpecialName,
 		DebuggerHidden,
@@ -122,6 +123,7 @@ namespace ICSharpCode.Decompiler.TypeSystem
 			new TopLevelTypeName("System.Runtime.CompilerServices", "NullablePublicOnlyAttribute"),
 			new TopLevelTypeName("System.Diagnostics", nameof(ConditionalAttribute)),
 			new TopLevelTypeName("System", nameof(ObsoleteAttribute)),
+			new TopLevelTypeName("Microsoft.CodeAnalysis", "EmbeddedAttribute"),
 			new TopLevelTypeName("System.Runtime.CompilerServices", "IsReadOnlyAttribute"),
 			new TopLevelTypeName("System.Runtime.CompilerServices", nameof(SpecialNameAttribute)),
 			new TopLevelTypeName("System.Diagnostics", nameof(DebuggerHiddenAttribute)),
